requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
authlib==1.7.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==47.0.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
joserfc==1.6.4
mako==1.3.11
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.0
pydantic-core==2.46.0
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.20.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pyyaml==6.0.3
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sqlalchemy==2.0.49
starlette==0.52.1
starlette-session-middleware==0.1.6
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
